backend/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
gunicorn==23.0.0
